fix: write tests for nostr-relaypool adapter * they are disabled for now since mocking isomorphic-ws doesn't work
