+++ packages/adapter-nostr-relaypool/tsconfig.json
@@ -9,4 +9,5 @@
     "noEmit": true,
   },
   "include": ["src/**/*.ts"],
+  "exclude": ["src/**/*.spec.ts"]
 }
